+++ plugin-bsp/magicmetamodel/src/main/kotlin/org/jetbrains/magicmetamodel/extensions/URIExtension.kt
@@ -1,0 +1,10 @@
+package org.jetbrains.magicmetamodel.extensions
+
+import java.net.URI
+import java.nio.file.Path
+import kotlin.io.path.toPath
+
+internal fun URI.toAbsolutePath(): Path =
+  this.normalize()
+    .toPath()
+    .toAbsolutePath()