+++ plugin-bsp/magicmetamodel/src/main/kotlin/org/jetbrains/magicmetamodel/impl/TargetsDetailsForDocumentProvider.kt
@@ -1,0 +1,83 @@
+package org.jetbrains.magicmetamodel.impl
+
+import ch.epfl.scala.bsp4j.BuildTargetIdentifier
+import ch.epfl.scala.bsp4j.SourceItem
+import ch.epfl.scala.bsp4j.SourcesItem
+import ch.epfl.scala.bsp4j.TextDocumentIdentifier
+import com.intellij.openapi.diagnostic.logger
+import com.intellij.openapi.diagnostic.trace
+import org.jetbrains.magicmetamodel.extensions.toAbsolutePath
+import java.net.URI
+import java.nio.file.Path
+import kotlin.reflect.KProperty
+
+internal class TargetsDetailsForDocumentProvider(sources: List<SourcesItem>) {
+
+  init {
+    LOGGER.trace { "Initializing TargetsDetailsForDocumentProvider..." }
+  }
+
+  private val documentIdToTargetsIdsMap by DocumentIdToTargetsIdsMapDelegate(sources)
+
+  init {
+    LOGGER.trace { "Initializing TargetsDetailsForDocumentProvider done!" }
+  }
+
+  fun getAllDocuments(): List<TextDocumentIdentifier> =
+    documentIdToTargetsIdsMap.keys
+      .map(this::mapPathToTextDocumentIdentifier)
+      .toList()
+
+  private fun mapPathToTextDocumentIdentifier(path: Path): TextDocumentIdentifier =
+    TextDocumentIdentifier(path.toUri().toString())
+
+  fun getTargetsDetailsForDocument(documentId: TextDocumentIdentifier): List<BuildTargetIdentifier> =
+    generateAllDocumentSubdirectoriesIncludingDocument(documentId)
+      .flatMap { documentIdToTargetsIdsMap[it] ?: emptyList() }
+      .toList()
+
+  private fun generateAllDocumentSubdirectoriesIncludingDocument(documentId: TextDocumentIdentifier): Sequence<Path> {
+    LOGGER.trace { "Generating all $documentId subdirectories..." }
+
+    val documentAbsolutePath = mapDocumentIdToAbsolutePath(documentId)
+
+    return generateSequence(documentAbsolutePath) { it.parent }
+      .also { LOGGER.trace { "Generating all $documentId subdirectories done! Subdirectories: $it." } }
+  }
+
+  private fun mapDocumentIdToAbsolutePath(documentId: TextDocumentIdentifier): Path =
+    URI.create(documentId.uri).toAbsolutePath()
+
+  companion object {
+    private val LOGGER = logger<TargetsDetailsForDocumentProvider>()
+  }
+}
+
+private class DocumentIdToTargetsIdsMapDelegate(private val sources: List<SourcesItem>) {
+
+  operator fun getValue(
+    thisRef: Any?,
+    property: KProperty<*>,
+  ): Map<Path, List<BuildTargetIdentifier>> {
+    LOGGER.trace { "Calculating document to target id map..." }
+
+    return sources
+      .flatMap(this::mapSourcesItemToPairsOfDocumentIdAndTargetId)
+      .groupBy({ it.first }, { it.second })
+      .also { LOGGER.trace { "Calculating document to target id map done! Map: $it." } }
+  }
+
+  private fun mapSourcesItemToPairsOfDocumentIdAndTargetId(
+    sourceItem: SourcesItem,
+  ): List<Pair<Path, BuildTargetIdentifier>> =
+    sourceItem.sources
+      .map(this::mapSourceItemToPath)
+      .map { Pair(it, sourceItem.target) }
+
+  private fun mapSourceItemToPath(sourceItem: SourceItem): Path =
+    URI.create(sourceItem.uri).toAbsolutePath()
+
+  companion object {
+    private val LOGGER = logger<DocumentIdToTargetsIdsMapDelegate>()
+  }
+}