+++ plugin-bsp/src/main/kotlin/org/jetbrains/plugins/bsp/listeners/MyProjectManagerListener.kt
@@ -1,13 +1,13 @@
-package com.github.abrams27.intellijbsp.listeners
+package org.jetbrains.plugins.bsp.listeners
 
-import com.github.abrams27.intellijbsp.services.MyProjectService
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ProjectManagerListener
+import org.jetbrains.plugins.bsp.services.MyProjectService
 
 internal class MyProjectManagerListener : ProjectManagerListener {
 
-    override fun projectOpened(project: Project) {
-        project.service<MyProjectService>()
-    }
+  override fun projectOpened(project: Project) {
+    project.service<MyProjectService>()
+  }
 }