+++ plugin-bsp/src/main/kotlin/org/jetbrains/plugins/bsp/services/MyProjectService.kt
@@ -1,0 +1,11 @@
+package org.jetbrains.plugins.bsp.services
+
+import com.intellij.openapi.project.Project
+import org.jetbrains.plugins.bsp.MyBundle
+
+class MyProjectService(project: Project) {
+
+  init {
+    println(MyBundle.message("projectService", project.name))
+  }
+}